pyproject.toml
[tool.poetry]
name = "git-commit-extractor"
version = "0.1.0"
description = "Extract commit info and diffs for a specific author"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
gitpython = "^3.1.40"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
